:truck: Move barCategorical to series directory
