+++ static/app/components/charts/series/barCategorical.tsx
